server/requirements.txt
flask
gunicorn
jsonschema
flask-cors
pymongo[srv]
Flask-JWT-Extended
redis
bleach
Pillow
boto3